+++ pytorch_lightning/loggers/__init__.py
@@ -26,7 +26,7 @@ __all__ = [
 
 from pytorch_lightning.loggers.comet import _COMET_AVAILABLE, CometLogger  # noqa: F401
 from pytorch_lightning.loggers.mlflow import _MLFLOW_AVAILABLE, MLFlowLogger  # noqa: F401
-from pytorch_lightning.loggers.neptune import _NEPTUNE_AVAILABLE, NeptuneLogger  # noqa: F401
+from pytorch_lightning.loggers.neptune import _NEPTUNE_AVAILABLE, NeptuneLegacyLogger  # noqa: F401
 from pytorch_lightning.loggers.test_tube import _TESTTUBE_AVAILABLE, TestTubeLogger  # noqa: F401
 from pytorch_lightning.loggers.wandb import _WANDB_AVAILABLE, WandbLogger  # noqa: F401
 
@@ -39,7 +39,7 @@ if _MLFLOW_AVAILABLE:
     __all__.append('MLFlowLogger')
 
 if _NEPTUNE_AVAILABLE:
-    __all__.append('NeptuneLogger')
+    __all__.append('NeptuneLegacyLogger')
 
 if _TESTTUBE_AVAILABLE:
     __all__.append('TestTubeLogger')